requirements.txt
tensorflow>=2.11.0
opencv-python-headless
face_recognition
streamlit
matplotlib
numpy
pandas
Pillow
